anvil: Move import_bitmap into anvil
